General: développements spécifiques Veolia : pas de reset au changement de site, ajout du zoom aux intents (beta)

+++ app/javascripts/factories/Smartgeo.js
@@ -76,6 +76,8 @@ angular.module('smartgeomobile').factory('Smartgeo', function($http, $window, $r
 
         _MAX_ID_FOR_SELECT_REQUEST : 4000,
 
+        _I_HAVE_SOME_SUICIDAL_TENDENCIES : true ,
+
         _intervals : {},
 
         /**
@@ -175,6 +177,9 @@ angular.module('smartgeomobile').factory('Smartgeo', function($http, $window, $r
          * Clear localStorage
          */
         reset: function(){
+            if(Smartgeo._I_HAVE_SOME_SUICIDAL_TENDENCIES){
+                return 'GUN_AGAINST_ME';
+            }
             localStorage.clear();
             var sites = Smartgeo.get_('sites') ;
             Smartgeo.unset_('sites');
@@ -557,10 +562,8 @@ angular.module('smartgeomobile').factory('Smartgeo', function($http, $window, $r
                         (error || function(){})();
                     });
             }, function(SqlError) {
-                console.log(SqlError.message);
+                console.log(SqlError);
                 console.log(request);
-                alertify.log(SqlError.message);
-                (error || function(){})();
             });
         },
